hardware/usart: Correct half-duplex mode.
Seems having both UDRE and TXC interrupts enabled is a real good way to
get things utterly confused.  Not an issue when running in full-duplex
and we're merely loading the next character in, but a real issue when
we're "done" transmitting and need to switch back to receive mode.

In short, we do *NOT* enable UDRE and TXC at the same time, otherwise
it's difficult to know if we're "done".  We enable the transmitter and
UDRE when we start transmitting, *THEN* when we hit the end of our
buffer, we switch to waiting for TXC only if our buffer is empty.

In TXC, we do one last check for data before finally switching over to
Rx mode.